rename pkg/utils to pkg/sync
On-behalf-of: @SAP christoph.mewes@sap.com
