+++ pkg/sync/resource_copy.go
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package utils
+package sync
 
 import (
 	"context"
