+++ pkg/sync/resource_copy_test.go
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package utils
+package sync
 
 import (
 	"testing"
